License fix: Move executor base classes to common
The `JobExecutor` interface is used extensively in dataset code
to run remote functions.
