+++ src/libertem/analysis/clust.py
@@ -6,7 +6,7 @@ from textwrap import dedent
 
 from .base import BaseAnalysis, AnalysisResult, AnalysisResultSet
 from libertem.common.async_utils import sync_to_async
-from libertem.executor.base import JobCancelledError
+from libertem.common.executor import JobCancelledError
 from libertem.udf.masks import ApplyMasksUDF
 from libertem.udf.stddev import StdDevUDF, consolidate_result
 from .getroi import get_roi
