+++ src/libertem/io/dataset/tvips.py
@@ -8,7 +8,7 @@ from glob import glob, escape
 
 from libertem.common.math import prod
 from libertem.common import Shape
-from libertem.executor.base import JobExecutor
+from libertem.common.executor import JobExecutor
 from libertem.common.messageconverter import MessageConverter
 from .base import (
     DataSet, DataSetException, DataSetMeta,
