+++ tests/executor/test_dask.py
@@ -8,7 +8,7 @@ import psutil
 from libertem.executor.dask import (
     CommonDaskMixin
 )
-from libertem.executor.scheduler import Worker, WorkerSet
+from libertem.common.scheduler import Worker, WorkerSet
 from libertem.executor.dask import DaskJobExecutor
 from libertem.executor.inline import InlineJobExecutor
 from libertem.udf.base import UDFRunner
